+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.4.6] - 2023-03-03
+
 ### Added
 
 - `WYSIWYGRegex` to `formElementsService`